Split class definition and implementation
This moves the implementation of member functions out
of the class. This will aid developers when using the
output the code can quickly be separated into headers
and cpp files.

The namespace is fully qualified in the class definition.
It is bad programming practice to put `using` keyword
in header files. By fully qualifying the names when
code is copy pasted into separate header files the using
keyword will not be needed.

Signed-off-by: George Nash <george.nash@intel.com>