Add in notification code
This adds in the code to handle when a notification
is sent to the observers. There are two variables that
handle starting and life time of the thread.

Three variables are used for controlling the when the
notification is sent. The two important variables are
the m_send_notification_flag and the condition_variable
m_cv.

When a value is changed via POST the m_send_notification_flag
is set to true. And m_cv.notify_all() is called to wake
the notification thread to send the notification.

The readme.md has been updated to indcate that the output now
supports notifications.

Signed-off-by: George Nash <george.nash@intel.com>