Add Platform class
This adds an additinal class to the output.
The Platfrom class is responsible for handling everything
associated with starting, registering platfrom info,
setting device info, and stopping the platform.

This causes all of the global variables that were being
used for platform info and device info to be moved into
a class reducing the amount of global variables.

All platform info can be accessed using the getPlatfromInfo
member function.  All device info is public member variables

Platform is no longer intialized by the IoTServer class.

Also made many small modifications to the template
file to clean up the white space of the output file.
most of these will be seen as adding a '-' to the jinja
if and for statments eather at the beinging and/or end of
the statments

Signed-off-by: George Nash <george.nash@intel.com>