Update resource class and PICS
The resource class no longer automatically registers
the resource. This was separated to make it possible
for the developer to catch and handle errors. Instead
of throwing a generic runtime error.

The resource constructor was also updated to accept
the resource URI as an input parameter. If no uri
is provided the code will still default to the uri
from the input json. This makes it possible to specify
multiple instances of the same resource on a device.

When registering the resource the developer has the option
of passing in the resource properties making is simpler to
change that value at the time of registration. If no value
is passed in when registering the resource it will default
to OC_DISCOVERABLE | OC_OBSERVABLE | OC_SECURE.

The resources oic.r.csr, oic.r.crl, and oic.r.roles was
added to the PICS file since these resources are currently
being added by default and the CTT will fail if the PICS
file does not contain those resources.

The PICS observableOICRES property was changes from true
to false. Testing has shown that /oic/res is currently
not observable. If we want /oic/res to be observable
the code must be updated.

Signed-off-by: George Nash <george.nash@intel.com>