NOTIFY code now uses notifyListOfObservers
The separate thread that used NotifyAllObservers was
removed and replaced with a call to notifyListOfObservers.
This removes almost all of the thread tracking code
that was used for NotifyAllObservers.

This change was made in responce to test results while
testing code against CTT.

Other small modifications. Documentation for output code
was updated. Removed comments that started with 'This function'
We know the documentation is for the function.

Also did some minor adjustments to spacing of the output.
Found some lines that should have had spacing while some other
lines needed the spacing removed.

Signed-off-by: George Nash <george.nash@intel.com>