Use compile time calculation for array iteration
This removes the separate variable that are used to track
the length of the arrays. Instead the sizeof operator is
used to calculate the interator value. This prevents
possible accessing past the end of the array if the
variable tracking the size of the array and actual array
size do not match.

Issue found using static code analysis tool.

Signed-off-by: George Nash <george.nash@intel.com>